First stab at different builds by wrapping component and injecting lottie.
